Fix comment and for loop commands and refactor commandPrefixes
