+++ src/definitions/codeSnippets.json
@@ -1,7 +1,7 @@
 [
   {
     "cmd": "add for loop",
-    "snippet": "for (let ${1:index} = 0; index < ${2:array}.length; ${1:index}++) {\n\t$3\n}\n"
+    "snippet": "for (let ${1:i} = ${2:0}; ${1:i} ${3:< array.length}; ${1:i}${4:++}) {\n\t$5\n}\n"
   },
   {
     "cmd": "add element container",
@@ -62,10 +62,6 @@
   {
     "cmd": "add new line",
     "snippet": "\n"
-  },
-  {
-    "cmd": "add comment",
-    "snippet": "// $1\n"
   },
   {
     "cmd": "add constant",
