+++ src/definitions/commandPrefixes.ts
@@ -38,14 +38,21 @@ export enum SystemPrefixes {
 export enum CompositionKeyword {
   for = "for",
   element = "element",
-  text = "text",
   else = "else",
   elseIf = "else if",
   indent = "indent",
   new = "new",
   keyword = "keyword",
-  symbol = "symbol",
   try = "try",
+}
+
+export enum CompositionTextKeyword {
+  text = "text",
+  comment = "comment",
+}
+
+export enum CompositionSymbolKeyword {
+  symbol = "symbol",
   condition = "condition",
 }
 
@@ -75,66 +82,3 @@ export enum EditingValue {
   anchor = "<a </a>",
   image = "<img />",
 }
-
-export const symbolsList = {
-  plus: "+",
-  minus: "-",
-  dash: "-",
-  star: "*",
-  times: "*",
-  curlyBraces: "{",
-  curlyBrackets: "{",
-  slash: "/",
-  dividedBy: "/",
-  angleBrackets: "<",
-  lessThan: "<",
-  lessThanOrEqualTo: "<=",
-  greaterThan: ">",
-  greaterThanOrEqualTo: ">=",
-  notEqual: "!=",
-  notEqualTo: "!=",
-  singleQuotes: "'",
-  doubleEqual: "==",
-  tripleEqual: "===",
-  equal: "=",
-  equalTo: "=",
-  roundBrackets: "(",
-  leftShift: "<<",
-  rightShift: ">>",
-  and: "&&",
-  squareBrackets: "[",
-  or: "||",
-  comma: ",",
-  colon: ":",
-  quotes: '"',
-  doubleQuotes: '"',
-  dot: ".",
-  period: ".",
-  underscore: "_",
-  semicolon: ";",
-  tripleQuotes: '"""',
-  bang: "!",
-  exclamation: "!",
-  questionMark: "?",
-  tilde: "~",
-  percent: "%",
-  at: "@",
-  dollar: "$",
-  doubleUnderscore: "__",
-  rightArrow: "->",
-  arrow: "->",
-  space: " ",
-  nullishCoalescing: "??",
-  backslash: "\\",
-  hash: "#",
-  caret: "^",
-  strictInequality: "!==",
-  ampersand: "&",
-  backtick: "`",
-  pipe: "|",
-  spread: "...",
-  leftBrace: "{",
-  rightBrace: "}",
-  leftBracket: "[",
-  rightBracket: "]",
-};
