+++ src/definitions/symbols.ts
@@ -1,0 +1,62 @@
+export const symbolsList = {
+    plus: "+",
+    minus: "-",
+    dash: "-",
+    star: "*",
+    times: "*",
+    curlyBraces: "{",
+    curlyBrackets: "{",
+    slash: "/",
+    dividedBy: "/",
+    angleBrackets: "<",
+    lessThan: "<",
+    lessThanOrEqualTo: "<=",
+    greaterThan: ">",
+    greaterThanOrEqualTo: ">=",
+    notEqual: "!=",
+    notEqualTo: "!=",
+    singleQuotes: "'",
+    doubleEqual: "==",
+    tripleEqual: "===",
+    equal: "=",
+    equalTo: "=",
+    roundBrackets: "(",
+    leftShift: "<<",
+    rightShift: ">>",
+    and: "&&",
+    squareBrackets: "[",
+    or: "||",
+    comma: ",",
+    colon: ":",
+    quotes: '"',
+    doubleQuotes: '"',
+    dot: ".",
+    period: ".",
+    underscore: "_",
+    semicolon: ";",
+    tripleQuotes: '"""',
+    bang: "!",
+    exclamation: "!",
+    questionMark: "?",
+    tilde: "~",
+    percent: "%",
+    at: "@",
+    dollar: "$",
+    doubleUnderscore: "__",
+    rightArrow: "->",
+    arrow: "->",
+    space: " ",
+    nullishCoalescing: "??",
+    backslash: "\\",
+    hash: "#",
+    caret: "^",
+    strictInequality: "!==",
+    ampersand: "&",
+    backtick: "`",
+    pipe: "|",
+    spread: "...",
+    leftBrace: "{",
+    rightBrace: "}",
+    leftBracket: "[",
+    rightBracket: "]",
+  };
